Selection-Sort
Selection sort => Space Complexity - O(1) and Time Complexity  - O(n^2)
Selection Sort is a sorting technique where the idea is to divide the array into two parts, sorted and unsorted.
We find the minimum element in the array and swap it with the left most element in the array, Now the leftmost element
becomes a part of sorted side and the rest of the array is unsorted. We repeat this Process until last index.

Selectionsort gives time complexity of O(n^2)  in all cases. The main work in Selectionsort is to find the index of 
the minimum element in the unsorted part of the array and push it to the leftmost index.