Merge-Sort
Merge-sort has a time-complexity of O(nlogn) in all cases. It is a divide and conquer technique and
also a comparison-based algorithm. Merge sort repeatedly breaks down a list into several sublists until each 
sublist consists of a single element and merging those sublists in a manner that results into a sorted list.'

Split the array into two, make a recursive call and merge the results. Divide the unsorted list into sublist of 
each one element, and repeatedly merge the sublists in a sorted manner until there is only one sublists remaining.